* Pull outer LandmarkComponent out of LandmarkDisplayScreenComponent
