+++ apps/mobile/feature/app/App.tsx
@@ -2,14 +2,14 @@ import { StatusBar } from 'expo-status-bar';
 import { useState } from 'react';
 import { SafeAreaView, StyleSheet } from 'react-native';
 
-import { LandmarkDisplayScreen } from './LandmarkDisplayScreen';
+import { LandmarkComponent } from './LandmarkComponent';
 import { SpeechService } from '../../service';
 
 export default function App() {
   const [speechService] = useState(new SpeechService());
   return (
     <SafeAreaView style={styles.container}>
-      <LandmarkDisplayScreen speechService={speechService} />
+      <LandmarkComponent speechService={speechService} />
       <StatusBar style="auto" />
     </SafeAreaView>
   );
